[config_mgmt.py]: Validate del ports before calling breakOutPorts API.
Changes:
-- Validate del ports before calling breakOutPorts API.
-- remove lib.py, since it is not needed now.
